+++ app.py
@@ -48,6 +48,7 @@ if not MANAGE_DB:
 
 @cache.memoize()
 def get_scenario_data(scenario_id):
+    app.logger.info(f"Loading scenario data #{scenario_id} (not cached)...")
     if USE_DUMMY_DATA:
         return dev.get_dummy_data(scenario_id)
     return scenario.get_scenario_data(scenario_id)
@@ -55,10 +56,13 @@ def get_scenario_data(scenario_id):
 
 @cache.memoize()
 def get_multiple_scenario_data(*scenario_ids):
+    app.logger.info("Merging scenario data (not cached)...")
     scenarios = [
         get_scenario_data(scenario_id) for scenario_id in scenario_ids
     ]
-    return scenario.merge_scenario_data(scenarios)
+    merged = scenario.merge_scenario_data(scenarios)
+    app.logger.info("Merged scenario data")
+    return merged
 
 
 @app.callback(
@@ -157,6 +161,7 @@ def load_scenario(scenarios):
         raise PreventUpdate
     scenarios = scenarios if isinstance(scenarios, list) else [scenarios]
     data = get_multiple_scenario_data(*scenarios)
+    app.logger.info("Data successfully loaded")
     return preprocessing.get_filter_options(data)